+++ tests/ControlsUnit/list_clean/BaseControl.test.ts
@@ -46,4 +46,47 @@ describe('Controls/list_clean/BaseControl', () => {
             assert.isTrue(!!baseControl._inertialScrolling);
         });
     });
+    describe('BaseControl watcher groupHistoryId', () => {
+
+        const GROUP_HISTORY_ID_NAME: string = 'MY_NEWS';
+
+        const baseControlCfg = {
+            viewName: 'Controls/List/ListView',
+            keyProperty: 'id',
+            viewModelConstructor: ListViewModel,
+            items: new RecordSet({
+                keyProperty: 'id',
+                rawData: []
+            })
+        };
+        let baseControl;
+
+        beforeEach(() => {
+            baseControl = new BaseControl(baseControlCfg);
+        });
+
+        afterEach(() => {
+            baseControl.destroy();
+            baseControl = undefined;
+        });
+
+        it('CollapsedGroup empty', () => {
+            baseControl._beforeMount(baseControlCfg);
+            baseControl._container = {getElementsByClassName: () => ([{clientHeight: 100, offsetHeight: 0}])};
+            baseControl._afterMount();
+            assert.isFalse(!!baseControl._listViewModel._options.collapsedGroups);
+        });
+        it('updated CollapsedGroups', async () => {
+            let cfgClone = {...baseControlCfg};
+            baseControl.saveOptions(baseControlCfg);
+            await baseControl._beforeMount(baseControlCfg);
+            baseControl._beforeUpdate(baseControlCfg);
+            baseControl._afterUpdate(baseControlCfg);
+            baseControl._container = {getElementsByClassName: () => ([{clientHeight: 100, offsetHeight: 0}])};
+            cfgClone.groupHistoryId = GROUP_HISTORY_ID_NAME;
+            cfgClone.collapsedGroups = [];
+            baseControl._beforeUpdate(cfgClone);
+            assert.isTrue(!!baseControl._listViewModel.getCollapsedGroups());
+        });
+    });
 });
