--- tests/ControlsUnit/list_clean/BaseController.test.ts
@@ -1,50 +1,0 @@
-import {assert} from 'chai';
-import {BaseControl, ListViewModel} from 'Controls/list';
-import {RecordSet} from 'Types/collection';
-
-describe('Controls/list_clean/BaseControl', () => {
-    // https://online.sbis.ru/opendoc.html?guid=9a6f0437-ea6d-4d7f-b163-25dc8f244c64 пункт 3
-    describe('BaseControl watcher groupHistoryId', () => {
-
-        const GROUP_HISTORY_ID_NAME: string = 'MY_NEWS';
-
-        const baseControlCfg = {
-            viewName: 'Controls/List/ListView',
-            keyProperty: 'id',
-            viewModelConstructor: ListViewModel,
-            items: new RecordSet({
-                keyProperty: 'id',
-                rawData: []
-            })
-        };
-        let baseControl;
-
-        beforeEach(() => {
-            baseControl = new BaseControl(baseControlCfg);
-        });
-
-        afterEach(() => {
-            baseControl.destroy();
-            baseControl = undefined;
-        });
-
-        it('CollapsedGroup empty', () => {
-            baseControl._beforeMount(baseControlCfg);
-            baseControl._container = {getElementsByClassName: () => ([{clientHeight: 100, offsetHeight: 0}])};
-            baseControl._afterMount();
-            assert.isFalse(!!baseControl._listViewModel._options.collapsedGroups);
-        });
-        it('updated CollapsedGroups', async () => {
-            let cfgClone = {...baseControlCfg};
-            baseControl.saveOptions(baseControlCfg);
-            await baseControl._beforeMount(baseControlCfg);
-            baseControl._beforeUpdate(baseControlCfg);
-            baseControl._afterUpdate(baseControlCfg);
-            baseControl._container = {getElementsByClassName: () => ([{clientHeight: 100, offsetHeight: 0}])};
-            cfgClone.groupHistoryId = GROUP_HISTORY_ID_NAME;
-            cfgClone.collapsedGroups = [];
-            baseControl._beforeUpdate(cfgClone);
-            assert.isTrue(!!baseControl._listViewModel.getCollapsedGroups());
-        });
-    });
-});
